// ws-client/src/app/school-erp/common/shared/model/education/education-type-entity.ts
export class EducationTypeEntity {


  oid: string;
  check: boolean;
  instituteOid: string;
  educationTypeOid: string;
  nameEn: string;
  nameBn: string;
  shortName: string;
  status: string;
  educationSystemOid: string;
  createdBy: string;
  educationClassLevelNameEn: string[];
  educationClassLevelNameBn: string[];
  educationGroupOid: string;
}