// ws-client/src/app/login/shared/model/user-info.ts
export class UserInfo {

  loginId: string;
  nameEn: string;
  nameBn: string;
  email: string;
  mobileNo: string;
  userPhotoPath: string;
  userPhotoUrl: string;
  instituteOid: string;
  status: string;
  roleType: string;
  studentId: string;
  teacherId: string;
  guardianId: string;
  scheduleMessagePush: string;
  principalId: string;
  accountantId: string;
  //roleOid: string;
  //menuJson: string;
  //loginOid: string;
  //loginStatus: string;
  //authorizationKey: string;
  //loginDateTime: string;
  districtOid: string;
  thanaOid: string;
}